+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file. The format 
 
 ## Unreleased
 
+- Added `centered` option to `Tabs` to center tab headings
 ## [0.22.2]
 - Improved visual bug involving skrinking / resizing for `Table` and `Accordion`